feat: factor out vm
